auth middleware

+++ controllers/stuffController.js
@@ -1,6 +1,6 @@
 const Thing = require("../models/thing");
 
-exports.createThing = (req, res, next) => {
+const createThing = (req, res, next) => {
   const thing = new Thing({
     title: req.body.title,
     description: req.body.description,
@@ -22,7 +22,7 @@ exports.createThing = (req, res, next) => {
     });
 };
 
-exports.getOneThing = (req, res, next) => {
+const getOneThing = (req, res, next) => {
   Thing.findOne({
     _id: req.params.id,
   })
@@ -36,7 +36,7 @@ exports.getOneThing = (req, res, next) => {
     });
 };
 
-exports.modifyThing = (req, res, next) => {
+const modifyThing = (req, res, next) => {
   const thing = new Thing({
     _id: req.params.id,
     title: req.body.title,
@@ -58,21 +58,33 @@ exports.modifyThing = (req, res, next) => {
     });
 };
 
-exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Deleted!",
+const deleteThing = (req, res, next) => {
+  Thing.findOne({ _id: req.params.id }).then((thing) => {
+    if (!thing) {
+      return res.status(404).json({
+        error: new Error("No such Thing"),
       });
-    })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
+    }
+    if (thing.userId !== req.auth.userId) {
+      return res
+        .status(401)
+        .json({ error: new Error("Unauthorized request!") });
+    }
+    Thing.deleteOne({ _id: req.params.id })
+      .then(() => {
+        res.status(200).json({
+          message: "Deleted!",
+        });
+      })
+      .catch((error) => {
+        res.status(400).json({
+          error: error,
+        });
       });
-    });
+  });
 };
 
-exports.getAllStuff = (req, res, next) => {
+const getAllStuff = (req, res, next) => {
   Thing.find()
     .then((things) => {
       res.status(200).json(things);
@@ -82,4 +94,12 @@ exports.getAllStuff = (req, res, next) => {
         error: error,
       });
     });
+};
+
+module.exports = {
+  createThing,
+  getAllStuff,
+  getOneThing,
+  deleteThing,
+  modifyThing,
 };
